Login Administracion funcionando

+++ Backend/administracion/src/main/java/ipn/escom/defensoria/administracion/entity/Rol.java
@@ -17,8 +17,7 @@ public class Rol {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private Integer id; 
     @Column(unique = true, nullable = false, length = 50)
     private String nombre;
 }
